Extract GotwayUnpacker into separate class, use constructor injection for GotwayAdapter
