+++ app/src/main/java/com/cooper/wheellog/utils/GotwayAdapter.java
@@ -6,13 +6,11 @@ import android.os.Handler;
 import com.cooper.wheellog.WheelData;
 import com.cooper.wheellog.WheelLog;
 
-import java.io.ByteArrayOutputStream;
-
 import timber.log.Timber;
 
 public class GotwayAdapter extends BaseAdapter {
     private static GotwayAdapter INSTANCE;
-    gotwayUnpacker unpacker = new gotwayUnpacker();
+    private GotwayUnpacker unpacker;
     private static final double RATIO_GW = 0.875;
     private String model = "";
     private String imu = "";
@@ -26,6 +24,12 @@ public class GotwayAdapter extends BaseAdapter {
     private final int alarmModeOne = 1; // 35 (45) km/h + 80% PWM
     private final int alarmModeOff = 2; // 80% PWM only
     private final int alarmModeCF = 3; // PWM tiltback for custom firmware
+
+    public GotwayAdapter(
+            GotwayUnpacker unpacker
+    ) {
+        this.unpacker = unpacker;
+    }
 
     @Override
     public boolean decode(byte[] data) {
@@ -355,54 +359,11 @@ public class GotwayAdapter extends BaseAdapter {
         }
     }
 
-    static class gotwayUnpacker {
-
-        enum UnpackerState {
-            unknown,
-            collecting,
-            done
-        }
-
-        ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-        gotwayUnpacker.UnpackerState state = UnpackerState.unknown;
-        int oldc = -1;
-
-        byte[] getBuffer() {
-            return buffer.toByteArray();
-        }
-
-        boolean addChar(int c) {
-            if (state == UnpackerState.collecting) {
-                buffer.write(c);
-                oldc = c;
-                int size = buffer.size();
-                if ((size == 20 && c != (byte) 0x18) || (size > 20 && size <= 24 && c != (byte) 0x5A)) {
-                    Timber.i("Invalid frame footer (expected 18 5A 5A 5A 5A)");
-                    state = UnpackerState.unknown;
-                    return false;
-                }
-                if (size == 24) {
-                    state = UnpackerState.done;
-                    Timber.i("Valid frame received");
-                    return true;
-                }
-            } else {
-                if (c == (byte) 0xAA && oldc == (byte) 0x55) {
-                    Timber.i("Frame header found (55 AA), collecting data");
-                    buffer = new ByteArrayOutputStream();
-                    buffer.write(0x55);
-                    buffer.write(0xAA);
-                    state = UnpackerState.collecting;
-                }
-                oldc = c;
-            }
-            return false;
-        }
-    }
-
     public static GotwayAdapter getInstance() {
         if (INSTANCE == null) {
-            INSTANCE = new GotwayAdapter();
+            INSTANCE = new GotwayAdapter(
+                    new GotwayUnpacker()
+            );
         }
         return INSTANCE;
     }
