+++ app/src/test/java/com/cooper/wheellog/RawDataTest.kt
@@ -39,7 +39,7 @@ class RawDataTest {
         every { config.gotwayNegative } returns "1"
         mockkConstructor(android.os.Handler::class)
         every { anyConstructed<android.os.Handler>().postDelayed(any(), any()) } returns true
-        val adapter = GotwayAdapter()
+        val adapter = GotwayAdapter(GotwayUnpacker())
         data.wheelType = Constants.WHEEL_TYPE.GOTWAY
         val inputStream: InputStream = File("src/test/resources/rawDecodeTest.csv").inputStream()
         val startTime = sdf.parse("11:50:50.123")
